go_dormant future and preemptible_repeat combinator
seastar threads have the advantage over plain futures
in that they can be made to have lesser impact on the CPU
by assigning them to a thread_scheduling_group.

this commit brings a similar functionality to futures.

first, the go_dormant(duration) future is similar to sleep,
but also gets awaken when cpu is idle.

second, the above is used in the preemptible_repeat
combinator which is similar to `repeat` but additionally
takes a pointer to thread_scheduling_group indicate
how the cpu usage shall be limited.

threads and futures share the same queue when going dormant
and thus they're awaken in a fifo order.

Signed-off-by: Peter Jankuliak <p.jankuliak@gmail.com>